imggen/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wfm-imggen"
version = "0.1.0"
description = "WiFi Marketing Image Generation Module — AI-powered marketing visuals via Google Gemini"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "wfm-shared",
    "google-genai>=1.0.0",
    "click>=8.0",
    "rich>=13.0",
    "gradio>=5.0",
    "Pillow>=10.0",
]

[tool.setuptools]
packages = ["imggen"]

[tool.setuptools.package-dir]
imggen = "."
[project.scripts]
imggen = "imggen.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["imggen"]
